extract signup footer to autthrecipemodule
